Added markdown Sanitizer.js
This function returns a converter object that has two plugins
registered on the postConversion plugin hook. One of them sanitizes
the output HTML to only include a list of whitelisted HTML tags.
The other one attempts to balance opening/closing tags to prevent
leaking formating if the HTML is displayed in the browser.